# perf(content-type parser): faster string collector